5 Acceptance testing

+++ app/controllers/questions_controller.rb
@@ -1,12 +1,16 @@
 # frozen_string_literal: true
 
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: %i[index show]
+
   expose :questions, -> { Question.all }
   expose :question
 
   def create
+    question.user = current_user
+
     if question.save
-      redirect_to question_path(question)
+      redirect_to question_path(question), notice: 'Your question was successfully created.'
     else
       render :new
     end
@@ -21,8 +25,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user&.author?(question)
+      question.destroy
+      redirect_to questions_path, notice: 'Your question was succesfully deleted.'
+    else
+      redirect_to question, alert: 'You are not the author of this question.'
+    end
   end
 
   private
